UltraCleaner: tidy the Windows build of the engine
Cross-compiling the engine with MinGW (the toolchain CI uses on both
Windows runners) showed four POSIX trash helpers built but unused there:
UniqueTrashName and MoveDirectoryOrFile are for the hand-rolled moves the
XDG and macOS paths do, and TrashInfoTimestamp and PercentEncodePath only
serve the .trashinfo record. Windows hands the whole job to the shell, so
guard them per platform family instead. That also makes
TrashInfoTimestamp's inner _WIN32 branch dead, so it drops to localtime_r.

Cast sizeof to DWORD for SHQUERYRBINFO::cbSize, and give the engine
target WIN32_LEAN_AND_MEAN and NOMINMAX — the same hygiene the core
library applies, which the engine does not inherit because it
deliberately does not link UltraCanvas.

No behaviour change on any platform.

+++ Apps/UltraCleaner/engine/UltraCleanerScanner.cpp
@@ -398,7 +398,7 @@ void Scanner::ScanDanglingSymlinks(const CleanRule& rule,
 void Scanner::ScanRecycleBin(const CleanRule& rule, ScanReport& report) {
 #if defined(_WIN32) || defined(_WIN64)
     SHQUERYRBINFO info{};
-    info.cbSize = sizeof(info);
+    info.cbSize = static_cast<DWORD>(sizeof(info));
     // A null root queries every drive's recycle bin at once.
     if (FAILED(SHQueryRecycleBinW(nullptr, &info))) return;
     if (info.i64NumItems == 0) return;
